Fix positioning bug with Modal
Modal is `position: fixed` but was positioned relative to the <body> tag
instead of the viewport, because the CSS Transform `perspective` was set
on the <body> tag.

Pushed it to the app DOM node so it is not a parent of the modal
and positioning works as expected.

Also adapted modal content top offset not to take mouse position
into account anymore, not needed with `position: fixed`
working correctly.