Fix click focus with action buttons.
If clicking very quickly on the icon to open the details modal,
the download button has not finished its transition and
is still located behind the "open details" icon.

Thanks to z-index, it was behind the "open details" icon but in front of
its link, taking focus, resulting in the download modal being opened
instead.

Fixed by removing z-index on the download button, so "open modal" link
takes focus when both are stacked.